requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
motor>=3.3.0
deepface>=0.0.80
opencv-python-headless>=4.8.0
numpy>=1.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pymongo[srv]>=4.6.0
requests>=2.31.0
